EZP-30979: Moved ComparisonResult into SPI
