+++ eZ/Publish/SPI/Comparison/ComparisonResult.php
@@ -6,7 +6,7 @@
  */
 declare(strict_types=1);
 
-namespace eZ\Publish\API\Repository\Values\Content\VersionDiff;
+namespace eZ\Publish\SPI\Comparison;
 
 interface ComparisonResult
 {
